Refactor and distinguish exports as components, types, etc.
